stb/requirements.txt
docx2txt==0.8
llama-index==0.10.37
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.37
llama-index-embeddings-azure-openai==0.1.9
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.8
llama-index-llms-openai==0.1.19
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.3
llamaindex-py-client==0.1.19
nltk==3.8.1
openai==1.30.1
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
tqdm==4.66.4

#docx2txt                                0.8
#llama-index                             0.10.37
#llama-index-agent-openai                0.2.5
#llama-index-cli                         0.1.12
#llama-index-core                        0.10.37
#llama-index-embeddings-azure-openai     0.1.9
#llama-index-embeddings-huggingface      0.2.0
#llama-index-embeddings-openai           0.1.9
#llama-index-indices-managed-llama-cloud 0.1.6
#llama-index-legacy                      0.9.48
#llama-index-llms-azure-openai           0.1.8
#llama-index-llms-openai                 0.1.19
#llama-index-multi-modal-llms-openai     0.1.6
#llama-index-program-openai              0.1.6
#llama-index-question-gen-openai         0.1.3
#llama-index-readers-file                0.1.22
#llama-index-readers-llama-parse         0.1.4
#llama-parse                             0.4.3
#llamaindex-py-client                    0.1.19
#nltk                                    3.8.1
#openai                                  1.30.1
#pydantic                                2.7.1
#pydantic_core                           2.18.2
#python-dotenv                           1.0.1
#tqdm                                    4.66.4
